requirements.txt
blessed==1.20.0
cffi==1.17.1
configparser==7.2.0
cryptography==44.0.2
editor==1.6.6
inquirer==3.4.0
pycparser==2.22
pyOpenSSL==25.0.0
readchar==4.2.1
runs==1.2.2
six==1.17.0
wcwidth==0.2.13
xmod==1.8.1
